feat: Add save and load functionality to AuroraData
This commit introduces the ability to save and load all read data in `AuroraData` to and from a JSON file.

- A new `AuroraDashboard.Data` class library has been created to hold the data model classes, improving modularity and testability.
- The `AuroraData` class now has `Save(string filePath)` and `Load(string filePath)` methods that use `System.Text.Json` for serialization.
- `JsonSerializerOptions` are configured to handle circular references and serialize public fields.
- A new test project, `AuroraDashboard.Tests`, has been added with tests to verify the save and load functionality.
