+++ AuroraDashboard.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 16.0.30621.155
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuroraDashboard", "AuroraDashboard\AuroraDashboard.csproj", "{07D29BAC-9CCD-4B28-853C-0DA656B93495}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuroraDashboard.Tests", "AuroraDashboard.Tests\AuroraDashboard.Tests.csproj", "{7436716D-898E-4615-BEBA-9419419ED085}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuroraDashboard.Data", "AuroraDashboard.Data\AuroraDashboard.Data.csproj", "{86C992C3-889B-406C-978C-A01F9654C758}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{07D29BAC-9CCD-4B28-853C-0DA656B93495}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07D29BAC-9CCD-4B28-853C-0DA656B93495}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07D29BAC-9CCD-4B28-853C-0DA656B93495}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7436716D-898E-4615-BEBA-9419419ED085}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7436716D-898E-4615-BEBA-9419419ED085}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7436716D-898E-4615-BEBA-9419419ED085}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7436716D-898E-4615-BEBA-9419419ED085}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86C992C3-889B-406C-978C-A01F9654C758}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86C992C3-889B-406C-978C-A01F9654C758}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86C992C3-889B-406C-978C-A01F9654C758}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86C992C3-889B-406C-978C-A01F9654C758}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
